refactor: Refactor exception
