--- PowerAutomateMockUp/FlowParser/CustomExceptions/FlowRunnerException.cs
@@ -1,11 +1,0 @@
-﻿using System;
-
-namespace Parser.FlowParser.CustomExceptions
-{
-    public class FlowRunnerException : Exception
-    {
-        public FlowRunnerException(string message) : base(message)
-        {
-        }
-    }
-}
